+++ reader/src/ASTR/menupage/search/storymenu.vue
@@ -26,7 +26,7 @@ export default {
         }
     },
     methods:{
-        loadStory(path){
+        loadStory(lang, path){
             var req = 's='+ this.server;
             req = req + '&f=' + path;
             window.location.search = req;